ai_tools_bridge/pyproject.toml
[project]
name = "ai-tools-bridge"
version = "0.1.0"
description = "FastAPI bridge server exposing all ai_tools_* packages over HTTP/JSON"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "pydantic>=2.12.0",
    "ai_tools_github",
    "ai_tools_jira",
    "ai_tools_confluence",
    "ai_tools_gerrit",
    "ai_tools_base",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/ai_tools_bridge"]

[dependency-groups]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=1.2.0",
    "httpx>=0.28.0",
    "ruff>=0.1.0",
]

[tool.uv.sources]
ai_tools_base = { path = "./packages/ai_tools_base", editable = true }
ai_tools_github = { path = "./packages/ai_tools_github", editable = true }
ai_tools_jira = { path = "./packages/ai_tools_jira", editable = true }
ai_tools_confluence = { path = "./packages/ai_tools_confluence", editable = true }
ai_tools_gerrit = { path = "./packages/ai_tools_gerrit", editable = true }

[[tool.uv.index]]
url = "https://pypi.org/simple"
name = "pypi"
default = true

[tool.ruff]
line-length = 120
target-version = "py312"
